refactor(orchestrator): split ssh and orchestrator into focused submodules
ssh.rs was a single 530-line file mixing three concerns. Split it into a
Rust-2024-style module (no mod.rs):
  ssh.rs            — SshConnectionManager public API
  ssh/command.rs    — CommandStatus, CommandContext
  ssh/connection.rs — SshConnection + russh internals

orchestrator.rs was a 490-line file covering four distinct phases. Split
into submodules with one impl block per file:
  orchestrator.rs              — Orchestrator struct, new, protocol, select_instances
  orchestrator/prepare.rs      — install, update, configure + ConfigureReport
  orchestrator/monitoring.rs   — start_monitoring + MonitoringReport
  orchestrator/run.rs          — cleanup, boot_nodes, run_nodes, run_clients,
                                  apply_faults_step, download_logs

Drop LocalGrafana from monitor.rs — it was dead code (allow(dead_code))
with no callers; remove the now-unused fs/PathBuf/MonitorError imports
alongside it.
